signup is ready to use

+++ Birahe/Models/Dto/PaymentDto's/Dto/PaymentVerifiedDto.cs
@@ -3,6 +3,6 @@ namespace Birahe.EndPoint.Models.Dto.PaymentDto_s.Dto;
 public class PaymentVerifiedDto {
     public string Username { get; set; }
     public string TeamName { get; set; }
-    public int? RefId { get; set; }
+    public Int128? RefId { get; set; }
     public int Amount { get; set; }
 }
