+++ Birahe/Models/Dto/PaymentDto's/ResponseModels/ZarinpalRequestData.cs
@@ -4,7 +4,7 @@ public class ZarinpalRequestData {
     public int code { get; set; }
     public string authority { get; set; } = null!;
 
-    public int fee { get; set; }
+    public Int64 fee { get; set; }
 
     public string fee_type { get; set; }
